- Fixed ldarga and conv.i1 test cases - Added more supported conversion operations (through AND) - Some combinations don't work and need fixes
git-svn-id: svn://localhost/mosa@161 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Test/Mosa/Runtime/CompilerFramework/IL/ConvI1.cs
@@ -38,7 +38,7 @@ namespace Test.Mosa.Runtime.CompilerFramework.IL
         public void ConvI1_I1(sbyte a)
         {
             CodeSource = "static class Test { static bool ConvI1_I1(sbyte expect, sbyte a) { return expect == (sbyte)a; } }";
-            Assert.IsTrue((bool)Run<Native_ConvI1_I1>("", "Test", "ConvI1_I1", ((sbyte)a), a));
+            Assert.IsTrue((bool)Run<Native_ConvI1_I1>("", "Test", "ConvI1_I1", a, a));
         }
 
         /// <summary>
